+++ miden-crypto/src/merkle/smt/large/storage/mod.rs
@@ -18,7 +18,10 @@ pub use error::{StorageError, StorageResult};
 #[cfg(feature = "rocksdb")]
 mod rocksdb;
 #[cfg(feature = "rocksdb")]
-pub use rocksdb::{RocksDbConfig, RocksDbSnapshotStorage, RocksDbStorage};
+pub use rocksdb::{
+    RocksDbBloomFilterBitsPerKey, RocksDbConfig, RocksDbDurabilityMode, RocksDbMemoryBudget,
+    RocksDbSnapshotStorage, RocksDbStorage, RocksDbTuningOptions, RocksDbWriteBufferManagerBudget,
+};
 
 mod memory;
 pub use memory::{MemoryStorage, MemoryStorageSnapshot};